show satellite detail and emmit position in every 3secs

+++ app/src/main/java/tr/com/mekhti/satellitesapp/data/model/Position.kt
@@ -1,7 +1,7 @@
 package tr.com.mekhti.satellitesapp.data.model
 
 class Position(
-    val posX: Float,
-    val posY: Float
+    val posX: Double,
+    val posY: Double
 ) {
 }
